+++ gmatlib/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0]
+### Added 
+- `Matrix<T>` `Sub` trait implementations 
+- `.iter()` method for `Matrix<T>`
+- `AddAssign` and `SubAssign` trait implementations.
+
+## [0.1.0]
+### Added 
+- `Matrix<T>` `Add` trait implementations 
+### Changed
+- Changed `extern "C"` function implementations to use `Box` instead of manual allocations
+
 ## [0.0.2]
 ### Added 
 - This `CHANGELOG.md` file.